Create README.md
This paper investigates the division of labour1 operative in naturally occurring bilingual discourse. It attempts to establish the grammar of intersentential codeswitching (ISC) in mixed Yoruba-English interrogative sentences. The paper seeks to find out what constitutes the asymmetry between the participating languages (Yoruba and English), as they supply important grammatical structures in order to produce a well-formed mixed YorubaEnglish interrogative sentences. Asymmetry in the division of labour is *We gratefully appreciate the invaluable comments of the anonymous reviewers on the previous versions of this paper. But for those comments, the improvement of this paper would not have been possible. This paper has been improved upon very much, but we doubt if it is up to the expectations of the reviewers. Special thanks to our very good friend (whom we’ve never met!), Ms Younha Lee, Editorial Assistant, Journal of Cognitive Science, Institute for Cognitive Science, Seoul National University, Korea (ROK) for her great assistance, understanding, patience and words of encouragement which gave the courage to go on, especially during the office fire incident when there seemed to be no hope. 1 This social science term was borrowed from Myers’-Scotton (2009: 81) To us, this metaphor beautifully describes what operates in codeswitching when two languages are used in a discourse to supply the morpho-syntactic structure.